# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/branches/{branchId}/theatres": {
            "post": {
                "description": "Create a new Theatre in a Branch",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin/Branch/Theatre"
                ],
                "summary": "Create Theatre",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Branch ID",
                        "name": "branchId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.TheatreResponse"
                        }
                    }
                }
            }
        },
        "/branches/{branchId}/theatres/{theatreId}": {
            "put": {
                "description": "Update a Theatre in a Branch by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin/Branch/Theatre"
                ],
                "summary": "Update Theatre",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Branch ID",
                        "name": "branchId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Theatre ID",
                        "name": "theatreId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TheatreResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a Theatre in a Branch by ID",
                "tags": [
                    "Admin/Branch/Theatre"
                ],
                "summary": "Delete Theatre",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Branch ID",
                        "name": "branchId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Theatre ID",
                        "name": "theatreId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
        "/customer/{customerId}/ticket/payments": {
            "get": {
                "description": "Get all payments",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Customer/Payment"
                ],
                "summary": "Get Payments",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Customer ID",
                        "name": "customerId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PaymentsResponse"
                        }
                    }
                }
            }
        },
        "/customer/{customerId}/ticket/payments/{paymentId}": {
            "get": {
                "description": "Get payment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Customer/Payment"
                ],
                "summary": "Get Payment",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Customer ID",
                        "name": "customerId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Customer ID",
                        "name": "paymentId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PaymentResponse"
                        }
                    }
                }
            }
        },
        "/customer/{customerId}/ticket/payments/{payment_id}": {
            "post": {
                "description": "Confirm payment with payment_id.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Customer/Payment"
                ],
                "summary": "Confirm Payment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "payment id",
                        "name": "payment_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PaymentResponse"
                        }
                    }
                }
            }
        },
...
        "/movies/{movieId}/schedule/{scheduleId}": {
            "get": {
                "description": "Get all schedules for a movie by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Customer/Schedule"
                ],
                "summary": "Get all schedules for a movie by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Movie ID",
                        "name": "movie_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ScheduleResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a schedule for a movie",
                "tags": [
                    "Admin/Schedule"
                ],
                "summary": "Update a movie schedule",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Movie ID",
                        "name": "movieId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Schedule ID",
                        "name": "scheduleId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "$ref": "#/definitions/models.ScheduleResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a schedule for a movie",
                "tags": [
                    "Admin/Schedule"
                ],
                "summary": "Create a movie schedule",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Movie ID",
                        "name": "movieId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Schedule ID",
                        "name": "scheduleId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.ScheduleResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a specific movie schedule by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin/Schedule"
                ],
                "summary": "Delete movie schedule",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Movie ID",
                        "name": "movie_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Schedule ID",
                        "name": "schedule_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/movies/{movieId}/schedules": {
            "get": {
                "description": "Get a schedule by movie_id and schedule_id.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Customer/Schedule"
                ],
                "summary": "Get Schedule",
                "parameters": [
                    {
                        "type": "string",
                        "description": "movie id",
                        "name": "movieId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SchedulesResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Branch": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "integer"
                },
                "theatres": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Theatre"
                    }
                }
            }
        },
...
        "models.Movie": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "rating": {
                    "type": "number"
                },
                "releaseDate": {
                    "type": "string"
                },
                "title": {
                    "type": "integer"
                }
            }
        },
...
        "models.Payment": {
            "type": "object",
            "properties": {
                "Amount": {
                    "type": "integer"
                },
                "ID": {
                    "type": "integer"
                },
                "Status": {
                    "$ref": "#/definitions/models.PaymentStatus"
                }
            }
        },
        "models.PaymentResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.Payment"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "models.PaymentStatus": {
            "type": "string",
            "enum": [
                "Pending",
                "Completed",
                "Expired"
            ],
            "x-enum-varnames": [
                "Pending",
                "Completed",
                "Expired"
            ]
        },
        "models.PaymentsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Payment"
                    }
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "models.Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "models.Schedule": {
            "type": "object",
            "properties": {
                "branch": {
                    "$ref": "#/definitions/models.Branch"
                },
                "id": {
                    "type": "integer"
                },
                "movie": {
                    "$ref": "#/definitions/models.Movie"
                },
                "price": {
                    "type": "integer"
                },
                "seats": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Seat"
                    }
                },
                "showtime": {
                    "type": "string"
                }
            }
        },
        "models.ScheduleResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.Schedule"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "models.SchedulesResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Schedule"
                    }
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "models.Seat": {
            "type": "object",
            "properties": {
                "availability": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "number": {
                    "type": "string"
                },
                "row": {
                    "type": "string"
                }
            }
        },
        "models.Theatre": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.TheatreResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.Theatre"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        }
    }
}